pyproject.toml
[tool.poetry]
name = "this-document-does-not-exists"
version = "0.1.0"
description = ""
authors = ["starminalush <alica154323@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
click = "^8.1.7"
pre-commit = "^4.0.1"
pyyaml = "^6.0.2"
faker = "^33.0.0"
pillow = "^11.0.0"
pandas = "^2.2.3"
docxtpl = "^0.19.0"
colorama = "^0.4.6"
requests = "^2.32.3"
pdoc3 = "^0.11.5"
pyment = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
src_paths = ["src"]

[tool.autoflake8]
max-line-length = 120
extend-ignore = ["D203", "E203", "E251", "E266", "E302", "E305", "E401", "E402", "E501", "F401", "F403", "W503"]
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10


[tool.flake8]
max-line-length = 120
extend-ignore = ["D203", "E203", "E251", "E266", "E302", "E305", "E401", "E402", "E501", "F401", "F403", "W503"]
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10
